requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1
python-dateutil==2.9.0.post0
gunicorn